Custom_Graph_RAG/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.33.0
neo4j==5.10.0
langchain==0.2.7
python-dotenv==1.0.0
